+++ src/QuartoNotebookWorker/ext/QuartoNotebookWorkerRemoteREPL.jl
@@ -27,4 +27,10 @@ function QuartoNotebookWorker._remote_repl(::Nothing, port)
     return new_port
 end
 
+function __init__()
+    if ccall(:jl_generating_output, Cint, ()) == 0
+        QuartoNotebookWorker.extension_loaded!(@__MODULE__)
+    end
+end
+
 end